requirements.txt
pandas
streamlit
plotly
fpdf2